requirements.txt
h5py==2.10
imageio==2.6.1
keras==2.2.4
matplotlib==3.1.1
numpy==1.18.5
pandas==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
tensorflow-gpu==1.14.0
keras==2.2.4
wget==3.2
